Break apart deprecate and import_alarm tests
